ui/ux improvements
