+++ mocktail-ui/src/components/editor/HeadersTab.tsx
@@ -1,0 +1,8 @@
+export function HeadersTab() {
+  return (
+    <div className="rounded-[9px] border border-dashed border-border p-3 text-[12.5px] text-muted">
+      Custom response headers land in <span className="font-semibold text-fg">v4</span> (new backend
+      field). Editor UI arrives with it.
+    </div>
+  )
+}
